// src/app/core/model/supervisor/supervisor.ts
export interface Supervisor {
    user_id: number;
    first_name: string;
    last_name: string;
    gender: number;
    mobile: string;
    mail: string;
    username: string;
    requests: number;
    sellers: number;
    status_id: number;
    created_at: Date;
    updated_at: Date;
    status_name: string;
}